lock root genome when indexing fragments too
