+++ src/edge/models/fragment.py
@@ -91,24 +91,12 @@ class Fragment(models.Model):
 
             return out_edges[0].to_chunk
 
-    def root(self):
-        if self.parent is None:
-            return self
-        return self.parent.root()
-
     def lock(self):
-        """
-        Do a select for update, which, when executed inside a transaction, places
-        a lock on the fragment.
-        """
-
-        root = self.root()
-
-        root_fragments = Fragment.objects.select_for_update().filter(pk=root.id)
-        # Lock only happens when queryset is evaluated, hence root_fragments[0]
-        root_fragment = root_fragments[0]
-        print(f"Lock fragment {root_fragment.id}")
-
+        genomes = [g for g in self.genome_set.all()]
+        if len(genomes) == 0:
+            return
+        genomes = sorted(genomes, key=lambda g: g.id)
+        genomes[0].lock()
         return self
 
     @transaction.atomic()
